:recycle: [Refactor] PostsManagerPage.tsx types 분리 #main
